pyproject.toml
[project]
name = "dms-system"
version = "0.1.0"
description = "Django Management System"
authors = [
    {name = "Mehdi"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
